chore: move database entities to own file
